LangChain/pyproject.toml
[project]
name = "jiri"
version = "0.1.0"
description = "Voice-first agentic bridge with MCP tool discovery"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "Jiri Team" }]

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=13.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    # Redis
    "redis>=5.2.0",
    # Configuration
    "pydantic>=2.10.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    # Logging
    "loguru>=0.7.2",
    # ChromaDB Client
    "chromadb>=0.4.22",
    # HTTP Client (for MCP discovery)
    "httpx>=0.28.0",
    # LLM
    "openai>=1.60.0",
    # LangGraph + MCP
    "langgraph>=0.4.0",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-openai>=0.3.0",
    "langchain>=0.3.0",
    "newsapi-python>=0.2.7",
    "mcp>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.8.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "black>=24.10.0",
    "httpx>=0.28.0",          # For TestClient
]

azure = ["azure-cognitiveservices-speech>=1.35.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py313"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # function call in default argument (needed for FastAPI Depends)
]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.isort]
known-first-party = ["jiri"]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["redis.*", "chromadb.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
